# src/main/java/com/odilo/interview/api/dto/ChangePasswordRequest.java
package com.odilo.interview.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ChangePasswordRequest {

    @NotBlank
    @Size(min = 6, max = 20, message = "Password size may contain 6 and 20 character")
    private String password;

}