# src/main/java/com/odilo/interview/api/dto/LoginResponse.java
package com.odilo.interview.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class LoginResponse {

    private String jwtToken;

}